Move all polling dispatching code into app/Jobs/DispatchPollJobs.php
